requirements.txt
flask
openai==0.28.0
PyPDF2
gunicorn
python-dotenv